feat: added toggle buy switch to expanded view
